Delete (now redundant) appcommons and appkommons module
