--- appcommons/build.gradle
@@ -1,12 +1,0 @@
-apply plugin: 'com.android.library'
-
-dependencies {
-    implementation project(":util")
-    api "androidx.appcompat:appcompat:$androidAppCompat"
-    api "com.google.dagger:dagger:$daggerVersion"
-    implementation 'com.jakewharton.timber:timber:4.7.1'
-
-    annotationProcessor "com.google.dagger:dagger-compiler:$daggerVersion"
-
-    testImplementation "junit:junit:$junit4"
-}
