--- appcommons/gradle.properties
@@ -1,3 +1,0 @@
-POM_NAME=Android appcommons
-POM_DESCRIPTION=Android appcommons
-POM_ARTIFACT_ID=appcommons
