--- appcommons/src/main/java/ng/kingsley/android/app/ShutdownCompatActivity.java
@@ -1,15 +1,0 @@
-package ng.kingsley.android.app;
-
-import android.app.Activity;
-import android.os.Bundle;
-
-import androidx.core.app.ActivityCompat;
-
-public class ShutdownCompatActivity extends Activity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        ActivityCompat.finishAffinity(this);
-    }
-}
