--- appkommons/build.gradle
@@ -1,14 +1,0 @@
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-
-dependencies {
-    api project(":appcommons")
-    implementation project(":util")
-    api "com.google.android.material:material:$googleMaterial"
-    api "androidx.cardview:cardview:$androidCardView"
-    api "androidx.core:core-ktx:$androidCore"
-    api "com.google.code.gson:gson:$gsonVersion"
-    api "org.jetbrains.kotlin:kotlin-stdlib:$kotlinVersion"
-
-    testImplementation "junit:junit:$junit4"
-}
