--- appkommons/gradle.properties
@@ -1,3 +1,0 @@
-POM_NAME=Android appkommons
-POM_DESCRIPTION=Android appkommons
-POM_ARTIFACT_ID=appkommons
