--- appkommons/src/main/AndroidManifest.xml
@@ -1,11 +1,0 @@
-<manifest package="ng.kingsley.android.appkommons"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <application
-        android:allowBackup="true"
-        android:label="@string/app_name"
-        android:supportsRtl="true">
-
-    </application>
-
-</manifest>
