--- appkommons/src/main/res/values/strings.xml
@@ -1,4 +1,0 @@
-<resources>
-    <string name="app_name">Appcommons Kotlin</string>
-    <string name="wait">Please wait…</string>
-</resources>
