--- appkommons/src/test/java/ng/kingsley/android/appkommons/ExampleUnitTest.java
@@ -1,15 +1,0 @@
-package ng.kingsley.android.appkommons;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * To work on unit tests, switch the Test Artifact in the Build Variants view.
- */
-public class ExampleUnitTest {
-    @Test
-    public void addition_isCorrect() throws Exception {
-        assertEquals(4, 2 + 2);
-    }
-}
