+++ bootstrap/gradle.properties
@@ -1,0 +1,3 @@
+POM_NAME=bootstrap
+POM_DESCRIPTION=Appcommons bootstrap
+POM_ARTIFACT_ID=bootstrap
