+++ bootstrap/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<resources>
+    <string name="app_name">Bootstrap</string>
+    <string name="wait">Please wait…</string>
+</resources>
