+++ settings.gradle
@@ -1,5 +1,3 @@
-include ':appcommons'
-include ':appkommons'
 include ':recyclerview-ext'
 include ':demo'
 include ':preference'
@@ -7,3 +5,4 @@ include ':preference-gson'
 include ':preference-moshi'
 include ':util'
 include ':view'
+include ':bootstrap'
