+++ util/src/main/java/com/kingsleyadio/appcommons/util/ContentResolver.kt
@@ -1,4 +1,4 @@
-package ng.kingsley.android.extensions
+package com.kingsleyadio.appcommons.util
 
 import android.content.ContentResolver
 import android.net.Uri
